backend/requirements.txt
fastapi
uvicorn
python-dotenv
azure-storage-blob
azure-ai-formrecognizer
azure-core
azure-cosmos
openai
fastapi
uvicorn
python-dotenv
azure-storage-blob
azure-ai-formrecognizer
openai
pydantic
